+++ src/main/java/vn/tingee/sdk/model/DownloadInvoiceQueryDto.java
@@ -12,24 +12,14 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public class DownloadInvoiceQueryDto {
     @JsonProperty("merchantId")
     private Integer merchantId;
-    @JsonProperty("provider")
-    private EInvoiceProviderEnum provider;
-    @JsonProperty("taxCode")
-    private String taxCode;
-    @lombok.NonNull
-    @JsonProperty("invoicePattern")
-    private String invoicePattern;
-    @lombok.NonNull
-    @JsonProperty("invoiceSeries")
-    private String invoiceSeries;
+    @JsonProperty("accountId")
+    private Integer accountId;
     @lombok.NonNull
     @JsonProperty("invoiceCode")
     private String invoiceCode;
 
     /** Use this constructor to set all required fields. */
-    public DownloadInvoiceQueryDto(String invoicePattern, String invoiceSeries, String invoiceCode) {
-        this.invoicePattern = invoicePattern;
-        this.invoiceSeries = invoiceSeries;
+    public DownloadInvoiceQueryDto(String invoiceCode) {
         this.invoiceCode = invoiceCode;
     }
 }